# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "UpcycleMatch API",
        "description": "API Documentation untuk UpcycleMatch. Mendukung JWT Bearer Token dan Basic Auth.",
        "contact": {
            "email": "admin@upcyclematch.id"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Server"
        }
    ],
    "paths": {
...
        "/api/v1/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Daftar Pesanan",
                "description": "Melihat daftar pesanan. Contributor melihat pesanannya, Upcycler melihat pesanan masuk.",
                "operationId": "04e0961368d6d76f2eda41b4130cbf0c",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data pesanan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Buat Pesanan Baru",
                "description": "Hanya untuk role Contributor. Membuat pesanan baru untuk membeli produk Upcycle.",
                "operationId": "323b1672f01bdf0c1abc5653ff4ef17e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_id",
                                    "quantity",
                                    "recipient_name",
                                    "recipient_phone",
                                    "address",
                                    "city",
                                    "payment_method"
                                ],
                                "properties": {
                                    "product_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "recipient_name": {
                                        "type": "string",
                                        "example": "Andi"
                                    },
                                    "recipient_phone": {
                                        "type": "string",
                                        "example": "08123456789"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Jl. Sudirman 10"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Jakarta"
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "example": "transfer",
                                        "enum": [
                                            "koin",
                                            "transfer"
                                        ]
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Tolong packing rapi"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pesanan berhasil dibuat"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/orders/{id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Detail Pesanan",
                "description": "Mengambil detail dari satu data pesanan berdasarkan ID.",
                "operationId": "2af619ca9a78b4fc6c7911cdcb9555e9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID pesanan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil detail pesanan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "basicAuth": {
                "type": "http",
                "scheme": "basic"
            },
            "apiKeyAuth": {
                "type": "apiKey",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    },
...
}